[#4340] introduce separate autoconfigurations for each converter implementation
for more clarity we split the converter autoconfigurations from the converter infrastructure autoconfigurations.
