+++ extensions/spring/spring-boot-autoconfigure/src/main/java/org/axonframework/extension/springboot/autoconfig/Jackson2MapperAutoConfiguration.java
@@ -17,29 +17,16 @@
 package org.axonframework.extension.springboot.autoconfig;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.axonframework.conversion.ChainingContentTypeConverter;
-import org.axonframework.conversion.Converter;
-import org.axonframework.conversion.jackson2.Jackson2Converter;
 import org.axonframework.extension.springboot.ConverterProperties;
-import org.axonframework.messaging.core.conversion.DelegatingMessageConverter;
-import org.axonframework.messaging.core.conversion.MessageConverter;
-import org.axonframework.messaging.eventhandling.conversion.DelegatingEventConverter;
-import org.axonframework.messaging.eventhandling.conversion.EventConverter;
-import org.springframework.beans.factory.BeanClassLoaderAware;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.AnyNestedCondition;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
-import org.springframework.context.annotation.Primary;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * Autoconfiguration that constructs a default {@link ObjectMapper}, typically to be used by a
@@ -49,93 +36,15 @@ import static java.util.Objects.requireNonNull;
  * @since 5.0.3
  */
 @AutoConfiguration
-@AutoConfigureBefore({AxonAutoConfiguration.class, CBORMapperAutoConfiguration.class})
+@AutoConfigureBefore({AxonAutoConfiguration.class,
+        CBORMapperAutoConfiguration.class,
+        Jackson2ConverterAutoConfiguration.class})
 @AutoConfigureAfter(name = {
         "org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration",
         "org.springframework.boot.jackson.autoconfigure.JacksonAutoConfiguration"
 })
 @ConditionalOnClass(name = "com.fasterxml.jackson.databind.ObjectMapper")
-@EnableConfigurationProperties(value = ConverterProperties.class)
-public class Jackson2MapperAutoConfiguration implements BeanClassLoaderAware {
-
-    private ClassLoader classLoader;
-
-    /**
-     * Bean creation method constructing a {@link Jackson2Converter} as the "general" {@link Converter} to be used by
-     * Axon Framework.
-     *
-     * @param objectMapper the {@link ObjectMapper} to be used
-     * @return the "general" {@link Converter} to be used by Axon Framework
-     */
-    @Bean
-    @Primary
-    @ConditionalOnMissingBean(ignored = {MessageConverter.class, EventConverter.class})
-    @ConditionalOnProperty(name = "axon.converter.general", havingValue = "jackson2")
-    public Converter converter(ObjectMapper objectMapper) {
-        return buildConverter(objectMapper);
-    }
-
-    /**
-     * Bean creation method constructing a {@link MessageConverter} delegating to the "general" {@link Converter} in
-     * case both use {@code jackson2}.
-     *
-     * @param generalConverter the "general" {@link Converter}, used to construct the {@link MessageConverter} in case
-     *                         both use {@code jackson2}
-     * @return the {@link MessageConverter} to be used by Axon Framework
-     */
-    @Bean(name = "messageConverter")
-    @ConditionalOnMissingBean
-    @ConditionalOnExpression("'${axon.converter.messages}' == 'jackson2' && '${axon.converter.general}' == 'jackson2'")
-    public MessageConverter delegatingMessageConverter(Converter generalConverter) {
-        return new DelegatingMessageConverter(generalConverter);
-    }
-
-    /**
-     * Bean creation method constructing a {@link Jackson2Converter} as the {@link MessageConverter} to be used by Axon
-     * Framework.
-     *
-     * @param objectMapper the {@link ObjectMapper} to be used
-     * @return the {@link MessageConverter} to be used by Axon Framework.
-     */
-    @Bean
-    @ConditionalOnMissingBean
-    @ConditionalOnExpression("'${axon.converter.messages}' == 'jackson2' && '${axon.converter.general}' != 'jackson2'")
-    public MessageConverter messageConverter(ObjectMapper objectMapper) {
-        return new DelegatingMessageConverter(buildConverter(objectMapper));
-    }
-
-    /**
-     * Bean creation method constructing an {@link EventConverter} delegating to the {@link MessageConverter} in case
-     * both use {@code jackson2}.
-     *
-     * @param messageConverter the {@link MessageConverter}, used to construct the {@link EventConverter} in case both
-     *                         use {@code jackson2}
-     * @return the {@link EventConverter} to be used by Axon Framework.
-     */
-    @Bean(name = "eventConverter")
-    @ConditionalOnMissingBean
-    @ConditionalOnExpression("'${axon.converter.events}' == 'jackson2' && '${axon.converter.messages}' == 'jackson2'")
-    public EventConverter delegatingEventConverter(MessageConverter messageConverter) {
-        return new DelegatingEventConverter(messageConverter);
-    }
-
-    /**
-     * Bean creation method constructing a {@link Jackson2Converter} as the {@link EventConverter} to be used by Axon
-     * Framework.
-     *
-     * @param objectMapper the {@link ObjectMapper} to be used
-     * @return the {@link EventConverter} to be used by Axon Framework
-     */
-    @Bean
-    @ConditionalOnMissingBean
-    @ConditionalOnExpression("'${axon.converter.events}' == 'jackson2' && '${axon.converter.messages}' != 'jackson2'")
-    public EventConverter eventConverter(ObjectMapper objectMapper) {
-        return new DelegatingEventConverter(buildConverter(objectMapper));
-    }
-
-    private Converter buildConverter(ObjectMapper objectMapper) {
-        return new Jackson2Converter(objectMapper, new ChainingContentTypeConverter(classLoader));
-    }
+public class Jackson2MapperAutoConfiguration {
 
     /**
      * Returns the default Axon Framework {@link ObjectMapper}, if required.
@@ -185,17 +94,5 @@ public class Jackson2MapperAutoConfiguration implements BeanClassLoaderAware {
         static class EventsJacksonCondition {
 
         }
-    }
-
-    /**
-     * Sets the class loader used by the {@link ChainingContentTypeConverter} to load
-     * {@link org.axonframework.conversion.ContentTypeConverter ContentTypeConverters}.
-     *
-     * @param classLoader The class loader used by the {@link ChainingContentTypeConverter} to load
-     *                    {@link org.axonframework.conversion.ContentTypeConverter ContentTypeConverters}.
-     */
-    @Override
-    public void setBeanClassLoader(ClassLoader classLoader) {
-        this.classLoader = requireNonNull(classLoader, "The ClassLoader cannot be null.");
     }
 }
